third_party/blink/web_tests/platform/generic/http/tests/security/frame-loading-via-document-write-expected.txt

CONSOLE ERROR: Not allowed to load local resource: abe.png


--------
Frame: 'topRow'
--------
This page was successfully loaded.
My protocol is http:
My referrer is NOT blank


--------
Frame: 'middleRow'
--------
Image NOT loaded.

--------
Frame: 'bottomRow'
--------
Image loaded.
